src/services/google_meet_service.py
import asyncio
import logging
from typing import Dict, Optional

from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow
from googleapiclient.discovery import build
from googleapiclient.errors import HttpError

logger = logging.getLogger(__name__)

class GoogleMeetService:
    """Google Meet meeting service implementation."""
    
    def __init__(self, config: Dict):
        """Initialize the Google Meet service with configuration."""
        self.config = config
        self.credentials_path = config['google_meet']['credentials_path']
        self.scopes = [
            'https://www.googleapis.com/auth/calendar',
            'https://www.googleapis.com/auth/calendar.events',
            'https://www.googleapis.com/auth/meetings.space.created'
        ]
        self.service = None
        self.current_meeting = None
        self.audio_stream = None
        
        logger.info("Google Meet service initialized")
    
    async def join_meeting(self, meeting_id: str) -> bool:
        """Join a Google Meet meeting."""
        try:
            # Initialize Google Meet client
            credentials = await self._get_credentials()
            self.service = build('calendar', 'v3', credentials=credentials)
            
            # Get meeting details
            meeting = await self._get_meeting_details(meeting_id)
            if not meeting:
                raise ValueError(f"Meeting not found: {meeting_id}")
            
            # Join the meeting
            self.current_meeting = await self._join_meeting_session(meeting)
            
            # Start audio capture
            await self._start_audio_capture()
            
            logger.info(f"Successfully joined Google Meet: {meeting_id}")
            return True
        except Exception as e:
            logger.error(f"Failed to join Google Meet: {str(e)}")
            return False
    
    async def leave_meeting(self) -> bool:
        """Leave the current Google Meet meeting."""
        try:
            if self.current_meeting:
                # Stop audio capture
                await self._stop_audio_capture()
                
                # Leave the meeting
                await self._leave_meeting_session()
                
                self.current_meeting = None
                self.audio_stream = None
                
                logger.info("Successfully left Google Meet")
                return True
            return False
        except Exception as e:
            logger.error(f"Failed to leave Google Meet: {str(e)}")
            return False
    
    async def _get_credentials(self) -> Credentials:
        """Get Google API credentials."""
        try:
            flow = InstalledAppFlow.from_client_secrets_file(
                self.credentials_path,
                self.scopes
            )
            credentials = flow.run_local_server(port=0)
            return credentials
        except Exception as e:
            logger.error(f"Failed to get credentials: {str(e)}")
            raise
    
    async def _get_meeting_details(self, meeting_id: str) -> Dict:
        """Get details of a Google Meet meeting."""
        try:
            event = await self.service.events().get(
                calendarId='primary',
                eventId=meeting_id,
                fields='id,summary,description,hangoutLink'
            ).execute()
            return event
        except HttpError as e:
            logger.error(f"Failed to get meeting details: {str(e)}")
            return None
    
    async def _join_meeting_session(self, meeting: Dict) -> Dict:
        """Join a Google Meet session."""
        try:
            # Create a meeting session
            session = await self.service.events().patch(
                calendarId='primary',
                eventId=meeting['id'],
                body={
                    'description': f"{meeting.get('description', '')}\nJoined by Meeting Assistant Bot"
                }
            ).execute()
            
            return session
        except HttpError as e:
            logger.error(f"Failed to join meeting session: {str(e)}")
            raise
    
    async def _leave_meeting_session(self) -> None:
        """Leave the current Google Meet session."""
        try:
            if self.current_meeting:
                await self.service.events().patch(
                    calendarId='primary',
                    eventId=self.current_meeting['id'],
                    body={
                        'description': f"{self.current_meeting.get('description', '')}\nLeft by Meeting Assistant Bot"
                    }
                ).execute()
        except HttpError as e:
            logger.error(f"Failed to leave meeting session: {str(e)}")
            raise
    
    async def _start_audio_capture(self) -> None:
        """Start capturing audio from the Google Meet meeting."""
        try:
            # Initialize audio capture
            self.audio_stream = await self._initialize_audio_stream()
            
            # Start audio capture loop
            asyncio.create_task(self._audio_capture_loop())
            
            logger.info("Started audio capture for Google Meet")
        except Exception as e:
            logger.error(f"Failed to start audio capture: {str(e)}")
            raise
    
    async def _stop_audio_capture(self) -> None:
        """Stop capturing audio from the Google Meet meeting."""
        try:
            if self.audio_stream:
                await self.audio_stream.stop()
                self.audio_stream = None
                logger.info("Stopped audio capture for Google Meet")
        except Exception as e:
            logger.error(f"Failed to stop audio capture: {str(e)}")
            raise
    
    async def _initialize_audio_stream(self):
        """Initialize the audio stream for Google Meet."""
        # This is a placeholder for the actual audio stream initialization
        # You would need to implement the specific Google Meet audio capture logic
        pass
    
    async def _audio_capture_loop(self) -> None:
        """Main loop for capturing audio data."""
        try:
            while self.audio_stream and self.audio_stream.is_active:
                # Read audio data
                audio_data = await self.audio_stream.read()
                
                # Process audio data
                if audio_data:
                    # Convert to bytes if needed
                    if isinstance(audio_data, str):
                        audio_data = audio_data.encode('utf-8')
                    
                    # Yield audio data for processing
                    yield audio_data
                
                # Small delay to prevent CPU overload
                await asyncio.sleep(0.1)
        except Exception as e:
            logger.error(f"Error in audio capture loop: {str(e)}")
            raise
    
    @property
    def meeting_id(self) -> Optional[str]:
        """Get the current meeting ID."""
        return self.current_meeting['id'] if self.current_meeting else None 