requirements.txt
python-teams-bot>=1.0.0
google-meet-api>=1.0.0
openai>=1.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
jira>=3.5.1
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5
azure-cognitiveservices-speech>=1.20.0 